requirements.txt
flask>=3.1.0
psycopg2-binary>=2.9.0